+++ app/src/main/java/com/example/vitorio/exemplominicursoenuncomp2017/modelo/Ponto.java
@@ -7,10 +7,10 @@ package com.example.vitorio.exemplominicursoenuncomp2017.modelo;
 public class Ponto {
 
     private String nome;
-    private float latitude, longitude;
+    private double latitude, longitude;
 
     public Ponto(){}
-    public Ponto(String nome, float latitude, float longitude) {
+    public Ponto(String nome, double latitude, double longitude) {
         this.nome = nome;
         this.latitude = latitude;
         this.longitude = longitude;
@@ -24,19 +24,19 @@ public class Ponto {
         this.nome = nome;
     }
 
-    public float getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(float latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public float getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(float longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
 }
